Move utils to own module
Also, disable coverage for messages while they're not being used
