+++ raiden/src/messages/index.ts
@@ -1,2 +1,3 @@
+/* istanbul ignore file */
 export * from './types';
 export { packMessage } from './encode';
